fixed by #BlueGalaxy
Are you referring to the Pumpkin King quests? Here is the npc.

Attached is an updated event config and quests associated with them. You'll need to adjust the dates if you want to enable the events. I haven't fully tested these out yet.
To activate the event mission, you need to go to the data/static_data/events_config directory and find the entry:
<event name="Pumpkin Cookie Inventory" start="2020-10-30T00:00:00+00:00" end="2020-11-07T00:00:00+00:00">
It is
<event name="Pumpkin Cookie Event" start="2020-10-30T00:00:00+00:00" end="2020-11-07T00:00:00+00:00">
Then, where end="2020 is, you can replace it with end="2050. This will extend the event end date to the year 2050. Be sure to save your changes after making this modification. and to activate the "Halloween" event animation, go to config\main\events.properties and choose number 2
gameserver.enable.decor = 2