fixed by #DainAvenger
redesigned the placeholder for the rewards page. now this is a more elegant option

now if we have problems with the reward window, we will simply change addOnTalkEvent to addOnTalkEndEvent. This way we won't have fake rewards.

make a backup file

It's still a crappy solution, but it's better than it is