return to spawn using navmesh
After losing aggro, npc will try to return back to its spawn point using navmesh instead of just running through the walls. If there is no valid path, it will still pass through obstacles. If npc is over 100m away from the spawn and loses aggro, it will be "teleported" back to spawn.
Credits:
qingqing - initial implementation. It was modified a bit to remove a couple of bugs.
Aion Reconstruction Project - handling for the cases when an npc decides to go back to spawn, but the return distance is too large.